+++ eras/babbage/test-suite/test/Test/Cardano/Ledger/Babbage/Serialisation/Tripping.hs
@@ -10,7 +10,6 @@ import Cardano.Ledger.Babbage (Babbage)
 import Cardano.Ledger.Babbage.Rules (BabbageUtxoPredFailure)
 import Cardano.Ledger.Block (Block)
 import Cardano.Ledger.Core
-import Cardano.Ledger.Shelley.TxAuxData (ShelleyTxAuxData)
 import Cardano.Protocol.TPraos.BHeader (BHeader)
 import Test.Cardano.Ledger.Babbage.Serialisation.Generators ()
 import Test.Cardano.Ledger.Binary.RoundTrip
@@ -26,7 +25,7 @@ tests =
     [ testProperty "babbage/Script" $
         roundTripAnnExpectation @(Script Babbage),
       testProperty "babbage/Metadata" $
-        roundTripAnnExpectation @(ShelleyTxAuxData Babbage),
+        roundTripAnnExpectation @(TxAuxData Babbage),
       testProperty "babbage/TxOut" $
         roundTripCborExpectation @(TxOut Babbage),
       testProperty "babbage/TxBody" $